Complete GitHub orchestration MVP tasks
Split cognitive_engine.rs into 4 modules (<400 LoC each):
- cognitive_engine_core.rs (389 lines) - Main orchestration engine
- cognitive_engine_planning.rs (331 lines) - Planning and adjustment
- cognitive_engine_pr.rs (184 lines) - PR creation via MCP tool
- cognitive_engine_verification.rs (206 lines) - Verification checks

Replace gh CLI with MCP tool for PR creation:
- Use GitHubPrCreateKit MCP tool instead of direct shell commands
- Add A2A event logging for pr_creation_started/success/failed
- Structured metadata for debugging and metrics

Add comprehensive unit tests:
- 10 tests for PR creation logic in test_pr_creation.rs
- Tests for title/body formatting, MCP params, URL extraction
- Integration test for full PR creation flow (marked #[ignore])

Improve test assertions:
- Enhanced auth_manager_test.rs with detailed error messages
- Clear expected vs actual formatting for debugging

Document router integration:
- Created docs/router-integration.md (400+ lines)
- Covers initialization, usage patterns, model selection
- Includes code examples and file references

Add metrics and alerting:
- A2A event logging for PR creation with structured metadata
- Events include timing, success/failure, and error details
- No external SaaS dependencies

Implement rate limiting:
- Integrated arkavo-protocol RateLimiter with GitHub API
- Hybrid approach: global rate limit + GitHub-specific retry
- Respects x-ratelimit-reset headers from GitHub

Make webhook_secret optional:
- Changed to Option<String> to support polling mode (98% of users)
- Webhook server mode validates secret is present
- Config command shows "Not configured (polling mode)" when missing
- Clear error message guides users to set secret or use polling
